micro: split tool call think rewriter
